shop_payment_service/requirements.txt
asgiref
Django
djangorestframework
Pillow
pytz
sqlparse
typing_extensions
pytest
pytest-django
requests
django-cors-headers
coreapi
coreschema
stripe
django-filter
httpie